Moby Patch:
  * basic_opcodes.ops and opcode_table replaced by core.ops

  * process_opfunc.pl replaced by Parrot/OpsFile.pm, Parrot/Op.pm,
    and ops2c.pl. These convert *.ops to *_ops.c and
    include/parrot/oplib/*_ops.h

  * New ops2pm.pl converts *.ops to Parrot/OpLib/*.pm

  * disassemble.pl modified to use

  * interpreter.c modified to use structures in core_ops.c instead
    of the old op_info and op_func stuff. We're not yet to the
    point of handling multiple oplibs at a time, but we are getting
    there...

    Fingerprinting stuff disabled for now. We used to go off
    opcode_table. We need to rethink this. I have some ideas to
    make it unnecessary, but it remains to be seen whether others
    will like the direction they take us.

    Termination condition for while(){} loop in runops is now
    (pc) instead of (*pc). Any op can cause a HALT by returning
    a NULL pointer. Ops end, bounds and trace all do this. So, now
    the 'end' op body is actually executed, although all it does
    is HALT.

  * Makefile.in modified to perform a test build using pbc2c.pl,
    which is the bytecode --> C compiler I've posted to the
    list before.

  * Op full names are now autogenerated. Op root names must not
    contain underscores, so some op renamings have happened.
    Such as clear_i becoming cleari.


git-svn-id: https://svn.parrot.org/parrot/trunk@317 d31e2699-5ff4-0310-a27c-f18f2fbe73fe

+++ src/memory.c
@@ -69,7 +69,7 @@ mem_allocate_aligned(INTVAL size) {
 */
 void *
 mem_sys_allocate(INTVAL size) {
-    return malloc(size);
+    return malloc((size_t)size);
 }
 
 /*=for api mem mem_setup_allocator
